unit test framework
usage:

make test_parser

or

make all, to run all the test files

The main program has a makefile rule to make a lib out of the source
files. The tests makefile using the produced lib/archive to compile
individual test files, they are each their own main, this is to avoid
having to fix broken tests that are not part of the problem being acivly
worked on.

WARNING: this will not activly recompile the lib on changes,
something better should be worked out. Either direcly call the main lib
makefile to check, or something.